Make integer/float narrowings explicit in Proof.cpp and Gpu.cpp
Silence the C4244/C4267/C4334 (and a double->int) MSVC warnings on these two
units by making the existing implicit narrowings explicit, rather than widening
any types. Rationale:

- Every site that warned was already performing the same truncation implicitly,
  so an explicit cast generates identical codegen: zero runtime cost. Widening
  the variables to u64/double to "fix" the warning would have been the only
  change that actually cost anything, so it was deliberately avoided to keep the
  hot path's narrow u32/float types intact.
- None of these sites are in the per-square inner loop. They are one-time weight
  setup (weight32/invWeight32/...), periodic logging/ETA, ROE/carry stats, and
  end-of-run residue cleanup (doDiv3) -- so there is no maintainability-vs-speed
  trade to make here.

Proof.cpp: add a residueBytes(E) helper for the repeated (E-1)/8+1 byte count;
use nWords(E) for the residue word count (equal to the prior E/32+1 for the odd
prime exponents this path handles); cast middles.size()/points.size() to u32;
use size_t shifts in the 1u<<power asserts; assert span fits u32 before the
u32 expExp2 (exponents > 2^32 are not yet supported).

Gpu.cpp: explicit float()/u32()/int() casts at the weight returns, ROE/carry
stat counts, IterationTimer, squareLoop return, getETA call args, ulps delta,
doDiv3 quotients, and the CERT SHA3 byte count.

Verified correctness-neutral: res64 and Z are bit-identical to the recorded
FP64 1K:5:256 / -prp 19999999 baseline at both the 2000 and 10000 iter
checkpoints; per-iter timing unchanged (within run-to-run noise).

+++ src/Gpu.cpp
@@ -71,19 +71,19 @@ double invWeightM1(u32 N, u64 E, u32 H, u32 line, u32 col, u32 rep) {
 double boundUnderOne(double x) { return std::min(x, nexttoward(1, 0)); }
 
 float weight32(u32 N, u64 E, u32 H, u32 line, u32 col, u32 rep) {
-  return exp2((double)(extra(N, E, kAt(H, line, col) + rep)) / N);
+  return float(exp2((double)(extra(N, E, kAt(H, line, col) + rep)) / N));
 }
 
 float invWeight32(u32 N, u64 E, u32 H, u32 line, u32 col, u32 rep) {
-  return exp2(-(double)(extra(N, E, kAt(H, line, col) + rep)) / N);
+  return float(exp2(-(double)(extra(N, E, kAt(H, line, col) + rep)) / N));
 }
 
 float weightM132(u32 N, u64 E, u32 H, u32 line, u32 col, u32 rep) {
-  return expm1(M_LN2 * (double)(extra(N, E, kAt(H, line, col) + rep)) / N);
+  return float(expm1(M_LN2 * (double)(extra(N, E, kAt(H, line, col) + rep)) / N));
 }
 
 float invWeightM132(u32 N, u64 E, u32 H, u32 line, u32 col, u32 rep) {
-  return expm1(M_LN2 * - (double)(extra(N, E, kAt(H, line, col) + rep)) / N);
+  return float(expm1(M_LN2 * - (double)(extra(N, E, kAt(H, line, col) + rep)) / N));
 }
 
 Weights genWeights(FFTConfig fft, u64 E, u32 W, u32 H, u32 nW, bool nvidiaGpu) {
@@ -471,7 +471,7 @@ RoeInfo roeStat(const vector<float>& roe) {
     sumRoe  += x;
     sum2Roe += x * x;
   }
-  u32 n = roe.size();
+  u32 n = u32(roe.size());
 
   double sdRoe = sqrt(n * sum2Roe - sumRoe * sumRoe) / n;
   double meanRoe = sumRoe / n;
@@ -487,7 +487,7 @@ public:
   explicit IterationTimer(u64 kStart) : kStart(kStart) { }
 
   float reset(u64 k) {
-    float secs = timer.reset();
+    float secs = float(timer.reset());
 
     u64 its = max(u64(1), k - kStart);
     kStart = k;
@@ -1674,7 +1674,7 @@ u32 Gpu::squareLoop(Buffer<Word>& out, Buffer<Word>& in, u64 from, u64 to, bool 
     square(out, (k==from) ? in : out, leadIn, leadOut, doTailMul3 && (k == to - 1));
     leadIn = leadOut;
   }
-  return to;
+  return u32(to);
 }
 
 bool Gpu::isEqual(Buffer<Word>& in1, Buffer<Word>& in2) {
@@ -1743,7 +1743,7 @@ static string makeLogStr(const string& status, u64 k, u64 res, float secsPerIt, 
   
   snprintf(buf, sizeof(buf), "%2s %9" PRIu64 " %016" PRIx64 " %s ETA %s; ",
            status.c_str(), k, res, /* k / float(nIters) * 100, */
-           formatSecsPerIter(secsPerIt).c_str(), getETA(k, nIters, secsPerIt).c_str());
+           formatSecsPerIter(secsPerIt).c_str(), getETA(u32(k), u32(nIters), secsPerIt).c_str());
   return buf;
 }
 
@@ -1767,7 +1767,7 @@ void Gpu::doBigLog(u64 k, u64 res, bool checkOK, float secsPerIt, u64 nIters, u3
 
   RoeInfo carryStats = readCarryStats();
   if (carryStats.N > 2) {
-    u32 m = ldexp(carryStats.max, 32);
+    u32 m = u32(ldexp(carryStats.max, 32));
     double z = carryStats.z();
     log("Carry: %x Z(%u)=%.1f\n", m, carryStats.N, z);
   }
@@ -1785,7 +1785,7 @@ int ulps(double a, double b) {
   u64 aa = as<u64>(a);
   u64 bb = as<u64>(b);
   bool sameSign = (aa >> 63) == (bb >> 63);
-  int delta = sameSign ? bb - aa : bb + aa;
+  int delta = int(sameSign ? bb - aa : bb + aa);
   return delta;
 }
 
@@ -1875,12 +1875,12 @@ static void doDiv3(u64 E, Words& words) {
   assert(topBits > 0 && topBits < 32);
   {
     u64 w = (u64(r) << topBits) + words.back();
-    words.back() = w / 3;
+    words.back() = u32(w / 3);
     r = w % 3;
   }
   for (auto it = words.rbegin() + 1, end = words.rend(); it != end; ++it) {
     u64 w = (u64(r) << 32) + *it;
-    *it = w / 3;
+    *it = u32(w / 3);
     r = w % 3;
   }
 }
@@ -2297,7 +2297,7 @@ PRPResult Gpu::isPrimePRP(const Task& task) {
       log("   %9" PRIu64 " %016" PRIx64 " %s\n", k, res, formatSecsPerIter(secsPerIt).c_str());
       RoeInfo carryStats = readCarryStats();
       if (carryStats.N) {
-        u32 m = ldexp(carryStats.max, 32);
+        u32 m = u32(ldexp(carryStats.max, 32));
         double z = carryStats.z();
         log("Carry: %x Z(%u)=%.1f\n", m, carryStats.N, z);
       }
@@ -2420,7 +2420,7 @@ LLResult Gpu::isPrimeLL(const Task& task) {
 
     float secsPerIt = iterationTimer.reset(k);
     queue.setSquareTime((int) (secsPerIt * 1'000'000));
-    log("%9" PRIu64 " %016" PRIx64 " %s ETA %s\n", k, res64, formatSecsPerIter(secsPerIt).c_str(), getETA(k, kEnd, secsPerIt).c_str());
+    log("%9" PRIu64 " %016" PRIx64 " %s ETA %s\n", k, res64, formatSecsPerIter(secsPerIt).c_str(), getETA(u32(k), u32(kEnd), secsPerIt).c_str());
 
     if (k >= kEnd) { return {isAllZero, res64}; }
 
@@ -2484,7 +2484,7 @@ array<u64, 4> Gpu::isCERT(const Task& task) {
 
     if (k >= kEnd) {
       fs::remove (fname);
-      return std::move(SHA3{}.update(data.data(), (E-1)/8+1)).finish();
+      return std::move(SHA3{}.update(data.data(), u32((E-1)/8+1))).finish();
     }
 
     if (doStop) { throw "stop requested"; }
